Consolidated banned_addresses into bans
* Banned_addresses code moved into bans folder.

* Updated documentation for getMaliciousScore and addAddress

* This commit rebases master's changes
