+++ index.js
@@ -16,7 +16,6 @@ function core(opts) {
   core.polls = require(path.join(__dirname, 'polls'));
   core.notifications = require(path.join(__dirname, 'notifications'));
   core.moderationLogs = require(path.join(__dirname, 'moderation_logs'));
-  core.bannedAddresses = require(path.join(__dirname, 'banned_addresses'));
   core.helper = require(path.join(__dirname, 'helper'));
   core.db = require(path.join(__dirname, 'db'));
   core.close = function() { pg.end(); };
